feat: cleanup the PR from feedback
