+++ datafusion/core/Cargo.toml
@@ -102,6 +102,7 @@ bigdecimal = "0.4.1"
 criterion = { version = "0.5", features = ["async_tokio"] }
 csv = "1.1.6"
 ctor = "0.2.0"
+datafusion-sqllogictest = { path = "../sqllogictest", features = ["postgres"] }
 doc-comment = "0.3"
 env_logger = "0.10"
 half = "2.2.1"
