--- datafusion/core/tests/sqllogictests/src/engines/output.rs
@@ -1,57 +1,0 @@
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-use sqllogictest::{ColumnType, DBOutput};
-
-#[derive(Debug, PartialEq, Eq, Clone)]
-pub enum DFColumnType {
-    Boolean,
-    DateTime,
-    Integer,
-    Float,
-    Text,
-    Timestamp,
-    Another,
-}
-
-impl ColumnType for DFColumnType {
-    fn from_char(value: char) -> Option<Self> {
-        match value {
-            'B' => Some(Self::Boolean),
-            'D' => Some(Self::DateTime),
-            'I' => Some(Self::Integer),
-            'P' => Some(Self::Timestamp),
-            'R' => Some(Self::Float),
-            'T' => Some(Self::Text),
-            _ => Some(Self::Another),
-        }
-    }
-
-    fn to_char(&self) -> char {
-        match self {
-            Self::Boolean => 'B',
-            Self::DateTime => 'D',
-            Self::Integer => 'I',
-            Self::Timestamp => 'P',
-            Self::Float => 'R',
-            Self::Text => 'T',
-            Self::Another => '?',
-        }
-    }
-}
-
-pub type DFOutput = DBOutput<DFColumnType>;
