+++ datafusion/core/tests/sqllogictests/test_files/information_schema_columns.slt
