+++ datafusion/sqllogictest/src/engines/conversion.rs
@@ -123,3 +123,20 @@ pub fn i128_to_str(value: i128, precision: &u8, scale: &i8) -> String {
 pub fn big_decimal_to_str(value: BigDecimal) -> String {
     value.round(12).normalized().to_string()
 }
+
+/// Converts a 128-bit decimal into a string using specified precision and scale.
+///
+/// # Arguments
+///
+/// * `value` - The 128-bit decimal to convert.
+///
+/// # Panics
+///
+/// Panics if the decimal is not valid.
+///
+/// # Returns
+///
+/// The string representation of the decimal.
+pub fn decimal_to_str(value: Decimal) -> String {
+    big_decimal_to_str(BigDecimal::from_str(&value.to_string()).unwrap())
+}
