+++ datafusion/sqllogictest/src/engines/datafusion_engine/runner.rs
@@ -14,36 +14,40 @@
 // KIND, either express or implied.  See the License for the
 // specific language governing permissions and limitations
 
-use std::{sync::Arc, time::Duration};
+use std::{time::Duration, path::PathBuf};
 
 use arrow::record_batch::RecordBatch;
 use async_trait::async_trait;
 use datafusion::prelude::SessionContext;
-use sqllogictest::{DBOutput, TestError};
+use sqllogictest::DBOutput;
+use log::info;
+
+use super::{error::Result, DFSqlLogicTestError};
 
 use crate::engines::output::{DFColumnType, DFOutput};
 
 use super::{convert_batches, convert_schema_to_types};
 
-pub struct DataFusionTestRunner {
-    context: Arc<SessionContext>,
+pub struct DataFusionTestEngine {
+    ctx: SessionContext,
+    relative_path: PathBuf,
 }
 
-impl DataFusionTestRunner {
-    pub fn new(context: Arc<SessionContext>) -> Self {
-        Self { context }
+impl DataFusionTestEngine {
+    pub fn new(context: SessionContext, relative_path: PathBuf) -> Self {
+        Self { ctx: context, relative_path }
     }
 }
 
 async fn run_query(
     ctx: &SessionContext,
     sql: impl Into<String>,
-) -> Result<DFOutput, TestError> {
-    let df = ctx.sql(sql.into().as_str()).await.unwrap();
+) -> Result<DFOutput> {
+    let df = ctx.sql(sql.into().as_str()).await?;
 
     let types = convert_schema_to_types(df.schema().fields().as_slice());
-    let results: Vec<RecordBatch> = df.collect().await.unwrap();
-    let rows = convert_batches(results).unwrap();
+    let results: Vec<RecordBatch> = df.collect().await?;
+    let rows = convert_batches(results)?;
 
     if rows.is_empty() && types.is_empty() {
         Ok(DBOutput::StatementComplete(0))
@@ -53,17 +57,22 @@ async fn run_query(
 }
 
 #[async_trait]
-impl sqllogictest::AsyncDB for DataFusionTestRunner {
-    type Error = TestError;
+impl sqllogictest::AsyncDB for DataFusionTestEngine {
+    type Error = DFSqlLogicTestError;
     type ColumnType = DFColumnType;
 
-    async fn run(&mut self, sql: &str) -> Result<DFOutput, TestError> {
-        run_query(&self.context, sql).await
+    async fn run(&mut self, sql: &str) -> Result<DFOutput> {
+        info!(
+            "[{}] Running query: \"{}\"",
+            self.relative_path.display(),
+            sql
+        );
+        run_query(&self.ctx, sql).await
     }
 
     /// Engine name of current database.
     fn engine_name(&self) -> &str {
-        "DataFusionTestRunner"
+        "DataFusionTestEngine"
     }
 
     /// [`Runner`] calls this function to perform sleep.
